Fix app crashing bug with tags collection view
The bug would occur when tags were added, all tags were deleted, and
then another tag was added. Adding that last tag would cause a
collection view layout problem which was related to the height
constraint of the collection view I was manipulating. When the
collection was empty I was setting the constraint constant to zero.
Once it was set to a non-zero value the layout issue would occur.